Delineate alaGridMean and alaSpeciesMean
alaSpeciesMean is a reiteration of the previous alaGridMean.

alaSpeciesMeans returns the mean values per species. alaGridMean (not functional yet) will return the means per grid cell.

To accomodate for varying inputs/layouts in the .csv ALA exports, the function has been amended to specify the first and end columns corresponding to the layers/climate variables.

Updated script documentation and comments.